+++ src/cpp/simplify.h
@@ -1,8 +1,10 @@
 #ifndef SIMPLIFY_H
 #define SIMPLIFY_H
 
+#include <stddef.h>
+
 extern "C" {
-    void simplify_symengine_cpp(const char* input, char* output, int max_len);
+    void simplify_symengine_cpp(const char* input, char* output, size_t max_len);
 }
 
 #endif
